# perf: optimize sequence encoding with list comprehension