@YuriiMotov 2025-07-18
This PR is a duplicate of https://github.com/fastapi/fastapi/pull/12452.

I tried testing the #12452 and i got a decrease in performance instead of increase. See [my comment](https://github.com/fastapi/fastapi/pull/12452#pullrequestreview-3000759067).
The author of #12452 hasn't replied yet.
@killerdevildog, could you take a look and address it?